requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
backoff==2.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Deprecated==1.3.1
exceptiongroup==1.3.1
fastapi==0.123.0
flake8==7.3.0
googleapis-common-protos==1.59.1
grpcio==1.76.0
h11==0.16.0
honeycomb-opentelemetry==0.2.3b0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib-metadata==6.11.0
iniconfig==2.3.0
lance-namespace==0.2.1
lance-namespace-urllib3-client==0.2.1
mccabe==0.7.0
nodeenv==1.9.1
numpy==2.2.6
opentelemetry-api==1.20.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-instrumentation-aws-lambda==0.41b0
opentelemetry-instrumentation-dbapi==0.41b0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-grpc==0.41b0
opentelemetry-instrumentation-jinja2==0.41b0
opentelemetry-instrumentation-logging==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-sqlite3==0.41b0
opentelemetry-instrumentation-urllib==0.41b0
opentelemetry-instrumentation-urllib3==0.41b0
opentelemetry-instrumentation-wsgi==0.41b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==25.0
pluggy==1.6.0
protobuf==4.25.8
pyarrow==22.0.0
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pylance==0.39.0
pyright==1.1.407
pytest==9.0.1
python-dateutil==2.9.0.post0
requests==2.32.5
requests-mock==1.12.1
six==1.17.0
starlette==0.50.0
thrift==0.22.0
tomli==2.3.0
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urequests==0.1.2
urllib3==2.5.0
uvicorn==0.38.0